+++ backend/src/main/java/com/hotpack/krocs/domain/auth/controller/AuthController.java
@@ -1,0 +1,48 @@
+package com.hotpack.krocs.domain.auth.controller;
+
+import com.hotpack.krocs.domain.auth.dto.UserSession;
+import com.hotpack.krocs.domain.auth.service.AuthService;
+import com.hotpack.krocs.domain.auth.service.RedisTokenService;
+import com.hotpack.krocs.global.common.response.ApiResponse;
+import com.hotpack.krocs.global.security.annotation.Login;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseCookie;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CookieValue;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/auth")
+@RequiredArgsConstructor
+public class AuthController {
+
+    private final RedisTokenService redisTokenService;
+    private final AuthService authService;
+
+    @GetMapping("/me")
+    public ResponseEntity<ApiResponse<UserSession>> me(@Login UserSession userSession) {
+        if (userSession == null) {
+            return ResponseEntity.status(401).body(ApiResponse.onFailure("GLOBAL401", "인증 실패"));
+        }
+        return ResponseEntity.ok(ApiResponse.success(userSession));
+    }
+
+    @PostMapping("/logout")
+    public ApiResponse<Void> logout(@CookieValue("AUTH-TOKEN") String token,
+        HttpServletResponse response) {
+
+        if (redisTokenService.existsToken(token)) {
+            redisTokenService.revokeToken(token);
+        }
+
+        ResponseCookie cookie = authService.expireAuthTokenCookie();
+        response.addHeader("Set-Cookie", cookie.toString());
+
+        return ApiResponse.success();
+    }
+}
+
